Youtube Summarizer/requirements.txt
streamlit
langchain-openai
langchain-core
langchain-community
youtube-transcript-api
yt-dlp
openai
python-dotenv
